+++ src/main/java/git/utp/primerproyecto/primerproyecto/web/controllers/HotelController.java
@@ -28,7 +28,7 @@ public class HotelController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<HotelDTO> getHotel(@PathVariable Integer id){
+    public ResponseEntity<HotelDTO> getHotel(@PathVariable Long id){
         return new ResponseEntity<>(hotelService.getHotel(id), HttpStatus.OK);
     }
 
@@ -38,13 +38,13 @@ public class HotelController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<HotelDTO> updateHotel(@PathVariable Integer id, @RequestBody HotelDTO hotelDTO){
+    public ResponseEntity<HotelDTO> updateHotel(@PathVariable Long id, @RequestBody HotelDTO hotelDTO){
         HotelDTO updatedHotel = hotelService.updateHotel(id, hotelDTO);
         return new ResponseEntity<>(updatedHotel, HttpStatus.ACCEPTED);
     }
 
     @DeleteMapping("/{id}")
-    public  ResponseEntity<DeleteResponse> deleteHotel(@PathVariable Integer id){
+    public  ResponseEntity<DeleteResponse> deleteHotel(@PathVariable Long id){
         DeleteResponse deleteResponse = hotelService.deleteHotel(id);
         return new ResponseEntity<>(deleteResponse,HttpStatus.ACCEPTED);
     }
